Correct BSD license header
Removes the following conflicting clause form the header:
"THIS HEADER MAY NOT BE EXTRACTED OR MODIFIED IN ANY WAY."

On the one hand, this additional clause is redundant because
the BSD license already states that it must remain and the
copyright notice must be kept.
On the other hand, the clause freezes the file header
and prohibits future contributors from extending the
copyright notice for their contributions. This additional
clause is not part of the official BSD 3-clause.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>